refactor : 테스트 디렉터리 정리
