+++ src/test/java/com/hhplus/concert_ticketing/application/queue/QueueUseCaseIntegrationTest.java
@@ -1,9 +1,7 @@
-package com.hhplus.concert_ticketing.application;
+package com.hhplus.concert_ticketing.application.queue;
 
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
